TgBot/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "3.12.6"
aiogram = "3.26.0"
python-dotenv = "1.2.2"
sqlalchemy = "2.0.48"
aiosqlite = "0.22.1"
